+++ Tests/scripts/configure_tests.py
@@ -974,20 +974,14 @@ def get_test_list(files_string, branch_name, two_before_ga_ver='0', conf=None, i
     if is_conf_json:
         tests = tests.union(get_test_from_conf(branch_name, conf))
 
-    if sample_tests:  # Choosing 3 random tests for infrastructure testing
-        print_warning('Collecting sample tests due to: {}'.format(','.join(sample_tests)))
-        tests = tests.union(
-            get_random_tests(tests_num=RANDOM_TESTS_NUM, conf=conf, id_set=id_set, server_version=two_before_ga_ver))
-
     if not tests:
+        tests = get_random_tests(tests_num=RANDOM_TESTS_NUM, conf=conf, id_set=id_set, server_version=two_before_ga_ver)
         if changed_common:
             print_warning('Adding 3 random tests due to: {}'.format(','.join(changed_common)))
-            tests = tests.union(get_random_tests(tests_num=RANDOM_TESTS_NUM, conf=conf, id_set=id_set,
-                                                 server_version=two_before_ga_ver))
+        elif sample_tests:  # Choosing 3 random tests for infrastructure testing
+            print_warning('Collecting sample tests due to: {}'.format(','.join(sample_tests)))
         else:
             print_warning("Running Sanity check only")
-            tests = get_random_tests(tests_num=RANDOM_TESTS_NUM, conf=conf, id_set=id_set,
-                                     server_version=two_before_ga_ver)
             tests.add('DocumentationTest')  # test with integration configured
             tests.add('TestCommonPython')  # test with no integration configured