+++ Packs/AutoFocus/Integrations/AutofocusV2/AutofocusV2_test.py
@@ -1,4 +1,9 @@
 import json
+import pytest
+import requests
+import requests_mock
+import sys
+
 
 IP_ADDRESS = '127.0.0.1'
 
@@ -35,7 +40,13 @@ IP_RES_JSON = {
             "tag_definition_status_id": 1,
             "count": 7692947,
             "lasthit": "2019-12-11 02:29:45",
-            "description": "The Upatre Trojan typically arrives as an e-mail attachment or through an e-mail with a link to the malware. Upatre's function is to download additional malware onto the system, in most cases when the malware was initially observed downloading the Dyre banking Trojan which then attempts to steal the users online banking credentials which may may be used for fraud.\n\nSince then, the operators have diversified, with Upatre frequently seen downloading other banking trojans.\n",  # noqa: E501
+            "description": "The Upatre Trojan typically arrives as an e-mail attachment or through"
+                           " an e-mail with a link to the malware. Upatre's function is to download additional"
+                           " malware onto the system, in most cases when the malware was initially observed"
+                           " downloading the Dyre banking Trojan which then attempts to steal the users online"
+                           " banking credentials which may may be used for fraud.\n\nSince then, the operators have"
+                           " diversified, with Upatre frequently seen downloading other banking trojans.\n",
+            # noqa: E501
             "customer_name": "Palo Alto Networks Unit42",
             "customer_industry": "High Tech",
             "upVotes": 2,
@@ -47,7 +58,7 @@ IP_RES_JSON = {
             "tagGroups": [
                 {
                     "tag_group_name": "Downloader",
-                    "description": "This type of malware secretly downloads malicious files from a remote server, then installs and executes the files."  # noqa: E501
+                    "description": "This type of malware secretly downloads malicious files from a remote server, then installs and executes the files."  # noqa: E501
                 }
             ],
             "aliases": [
@@ -81,7 +92,13 @@ INDICATOR_RES = {
             "TagClassID": 3,
             "Count": 7692947,
             "Lasthit": "2019-12-11 02:29:45",
-            "Description": "The Upatre Trojan typically arrives as an e-mail attachment or through an e-mail with a link to the malware. Upatre's function is to download additional malware onto the system, in most cases when the malware was initially observed downloading the Dyre banking Trojan which then attempts to steal the users online banking credentials which may may be used for fraud.\n\nSince then, the operators have diversified, with Upatre frequently seen downloading other banking trojans.\n"  # noqa: E501
+            "Description": "The Upatre Trojan typically arrives as an e-mail attachment or through an e-mail with a"
+                           " link to the malware. Upatre's function is to download additional malware onto the system,"
+                           " in most cases when the malware was initially observed downloading the Dyre banking Trojan"
+                           " which then attempts to steal the users online banking credentials which may may be used "
+                           "for fraud.\n\nSince then, the operators have diversified, with Upatre frequently"
+                           " seen downloading other banking trojans.\n"
+            # noqa: E501
         }
     ]
 }
@@ -89,34 +106,39 @@ INDICATOR_RES = {
 
 def test_parse_indicator_response():
     from AutofocusV2 import parse_indicator_response
-    indicator = parse_indicator_response(IP_RES_JSON, 'IP')
+    raw_indicator = IP_RES_JSON['indicator']
+    raw_tags = IP_RES_JSON['tags']
+    indicator = parse_indicator_response(raw_indicator, raw_tags, 'IP')
     assert json.dumps(indicator) == json.dumps(INDICATOR_RES)
 
 
 def test_calculate_dbot_score():
     from AutofocusV2 import calculate_dbot_score
-    score = calculate_dbot_score(IP_RES_JSON, 'IP')
+    raw_indicator = IP_RES_JSON['indicator']
+    score = calculate_dbot_score(raw_indicator, 'IP')
     assert score == 3
 
 
 def test_calculate_dbot_score_file():
     from AutofocusV2 import calculate_dbot_score
-    score = calculate_dbot_score(FILE_RES_JSON, 'File')
+    raw_indicator = IP_RES_JSON['indicator']
+    score = calculate_dbot_score(raw_indicator, 'File')
     assert score == 3
 
 
-def test_get_indicator_outputs(mocker):
-    from AutofocusV2 import get_indicator_outputs
-    return_outputs_mock = mocker.patch('AutofocusV2.return_outputs')
+def test_connection_error(mocker):
 
-    indicator = [{'raw_response': IP_RES_JSON, 'value': IP_ADDRESS, 'score': 3, 'response': INDICATOR_RES}]
-    get_indicator_outputs('IP', indicator, 'Address')
-    outputs = return_outputs_mock.call_args[1]['outputs']
+    import AutofocusV2
 
-    assert return_outputs_mock.call_count == 1
-    assert outputs['DBotScore'][0]['Indicator'] == IP_ADDRESS
-    assert outputs['DBotScore'][0]['Score'] == 3
-    assert outputs['IP(val.Address && val.Address == obj.Address)'][0]['Address'] == IP_ADDRESS
-    assert outputs['IP(val.Address && val.Address == obj.Address)'][0]['Malicious']['Vendor'] == 'AutoFocus V2'
-    assert outputs['AutoFocus.IP(val.IndicatorValue === obj.IndicatorValue)'][0]['IndicatorValue'] == IP_ADDRESS
-    assert outputs['AutoFocus.IP(val.IndicatorValue === obj.IndicatorValue)'][0]['Tags'][0]['TagName'] == 'Upatre'
+    RETURN_ERROR_TARGET = 'AutofocusV2.return_error'
+    BASE_URL = 'https://autofocus.paloaltonetworks.com/api/v1.0'
+
+    return_error_mock = mocker.patch(RETURN_ERROR_TARGET, side_effect=sys.exit)
+
+    with requests_mock.Mocker() as m:
+        m.get(f'{BASE_URL}/tic', exc=requests.exceptions.ConnectionError)
+
+        with pytest.raises(SystemExit):
+            AutofocusV2.search_indicator('ip', '8.8.8.8')
+        assert 'Error connecting to server. Check your URL/Proxy/Certificate settings'\
+               in return_error_mock.call_args[0][0]